+++ tests/basic/if3.c
@@ -2,14 +2,14 @@
 #include <stdlib.h>
 #include <stdbool.h>
 
-int sjf_global();
+int32_t sjf_global();
 
-int sjf_global() {
-    int ifResult1;
+int32_t sjf_global() {
+    int32_t ifResult1;
     if (false) {
         ifResult1 = 1;
     } else {
-        ifResult1 = 0;
+        ifResult1 = (int32_t)0;
     }
     return ifResult1;
 }
